refactor: reactionCount 타입 변경 및 필요없는 변수 삭제

+++ android/app/src/main/java/com/app/edonymyeon/presentation/uimodel/ReactionCountUiModel.kt
@@ -6,5 +6,5 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class ReactionCountUiModel(
     val viewCount: Int,
-    val commentCount: Int,
+    val commentCount: Long,
 ) : Parcelable
